added plan to incorporate KEGG data
