+++ ligify/genbank/create_genbank.py
@@ -5,6 +5,18 @@ from genbank.annotations import get_plasmid_components
 from genbank.codon_optimize import codon_opt
 
 
+def create_plasmid(regulators, chemical):
+    for regulator in regulators:
+        result = create_genbank(
+            regulator["refseq"],
+            chemical,
+            regulator["protein"]["context"]["promoter"]["regulated_seq"],
+            regulator["reg_protein_seq"],
+        )
+        regulator["plasmid_sequence"] = str(result)
+
+    return regulators
+
 def create_genbank(regulator_name, ligand_name, promoter_seq, regulator_protein_seq):
     # Codon optimize the natural sequence
     opt_regulator_seq = codon_opt(regulator_protein_seq)
